[Clean] Add "operator" suffix to operator file names
- Add "operator" suffix to operator file names
