+++ src/operators/nullary-operator.test.ts
@@ -1,12 +1,12 @@
 // =============================================================================================================================
-// SRC - OPERATORS - EMPTY TEST
+// SRC - OPERATORS - NULLARY OPERATOR TEST
 // =============================================================================================================================
 // tslint:disable:only-arrow-functions no-unused-expression no-null-keyword
 import { expect } from "chai";
 import * as sinon from "sinon";
-import { createEmptyOperator } from "./empty";
+import { createNullaryOperator } from "./nullary-operator";
 
-describe("[ Empty Operator ]", function() {
+describe("[ Nullary Operator ]", function() {
   afterEach(function() {
     sinon.restore();
   });
@@ -15,22 +15,22 @@ describe("[ Empty Operator ]", function() {
   // Name
   // ---------------------------------------------------------------------------------------------------------------------------
   describe("NAME", function() {
-    it('should be "empty"', function() {
-      const name = createEmptyOperator().name;
+    it('should be "nullary"', function() {
+      const name = createNullaryOperator().name;
 
-      expect(name).to.eq("empty");
+      expect(name).to.eq("nullary");
     });
   });
 
   // ---------------------------------------------------------------------------------------------------------------------------
-  // Empty Rule
+  // Nullary Rule
   // ---------------------------------------------------------------------------------------------------------------------------
-  describe("EMPTY RULE", function() {
-    it('should be "isEmpty" checker', function() {
-      const isEmpty = sinon.spy();
-      const validators = createEmptyOperator({ isEmpty }).createValidators();
+  describe("NULLARY RULE", function() {
+    it('should be "isNullary" checker', function() {
+      const isNullary = sinon.spy();
+      const validators = createNullaryOperator({ isNullary }).createValidators();
 
-      expect(validators.empty).to.eq(isEmpty);
+      expect(validators.nullary).to.eq(isNullary);
     });
   });
 });
