+++ src/operators/number-operator.ts
@@ -1,5 +1,5 @@
 // =============================================================================================================================
-// SRC - OPERATORS - NUMBER
+// SRC - OPERATORS - NUMBER OPERATOR
 // =============================================================================================================================
 import {
   hasDigits,
