+++ src/operators/present-operator.ts
@@ -1,27 +1,27 @@
 // =============================================================================================================================
-// SRC - OPERATORS - EMPTY
+// SRC - OPERATORS - PRESENT OPERATOR
 // =============================================================================================================================
-import { isEmpty } from "@checkers/index";
+import { isPresent } from "@checkers/index";
 import { Validators } from "./shared";
 
 const deafultCheckers = {
-  isEmpty,
+  isPresent,
 };
 
 export interface Options {}
 
 type DICheckers = { [K in keyof typeof deafultCheckers]: typeof deafultCheckers[K] };
-export const createEmptyOperator = (checkers: Partial<DICheckers> = {}) => {
+export const createPresentOperator = (checkers: Partial<DICheckers> = {}) => {
   // tslint:disable:no-shadowed-variable
-  const { isEmpty } = { ...deafultCheckers, ...checkers };
+  const { isPresent } = { ...deafultCheckers, ...checkers };
   // tslint:enable:no-shadowed-variable
 
   return {
-    name: "empty",
+    name: "present",
     createValidators: (_: Options = {}) => {
       const validators: Validators = {};
 
-      validators.empty = isEmpty;
+      validators.present = isPresent;
 
       return validators;
     },
