+++ src/operators/string-operator.ts
@@ -1,5 +1,5 @@
 // =============================================================================================================================
-// SRC - OPERATORS - STRING
+// SRC - OPERATORS - STRING OPERATOR
 // =============================================================================================================================
 import {
   endsWith as checkToEndsWith,
